Issue#routes

+++ src/app/api/expenditure/[id].ts
@@ -1,11 +1,17 @@
+import { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../prisma/client";
+import { ExpenditureInput } from "@/app/constants/backend";
 
-export default async function handler(req, res) {
-  const { id } = req.query;
+export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 
   try {
     if (req.method === "GET") {
       // Fetch expenditure by ID
+      const id = req.query["id"] as string;
+
+      if (!id) {
+        return res.status(400).json({ error: "Project ID is required" });
+      }
       const expenditure = await prisma.expenditure.findUnique({
         where: { eid: id },
         include: {
@@ -21,23 +27,29 @@ export default async function handler(req, res) {
       return res.status(200).json(expenditure);
     } else if (req.method === "PUT") {
       // Update expenditure by ID
-      const { image, desc, name, amount, project, scheme } = req.body;
+      const { desc, name, amount } = req.body as ExpenditureInput;
+      const id = req.query["id"] as string;
 
+      if (!id) {
+        return res.status(400).json({ error: "Project ID is required" });
+      }
       const updatedExpenditure = await prisma.expenditure.update({
         where: { eid: id },
         data: {
-          image,
           desc,
           name,
           amount,
-          project: project ? { connect: { pid: project } } : undefined,
-          scheme: scheme ? { connect: { sid: scheme } } : undefined,
         },
       });
 
       return res.status(200).json(updatedExpenditure);
     } else if (req.method === "DELETE") {
       // Delete expenditure by ID
+      const id = req.query["id"] as string;
+
+      if (!id) {
+        return res.status(400).json({ error: "Project ID is required" });
+      }
       await prisma.expenditure.delete({
         where: { eid: id },
       });
